Change core request abstraction to use streams rather than byte arrays
This improves symmetry with Ring and enables better memory usage
characteristics. This is likely to be important when dealing with
large entities or large numbers of concurrent requests.

Happily, this shouldn't change the api of anything but the "core"
namespace - the input/output coercers have been adjusted to preserve
the existing API.

Note that I'm not sure I got the content length logic right - more
real world testing would be useful.